Create an auth object on your request object...

+++ backend/controllers/sauce.js
@@ -73,18 +73,32 @@ exports.modifySauce = (req, res, next) => {
 };
 
 exports.deleteSauce = (req, res, next) => {
-  Sauce.deleteOne({ _id: req.parms.id }).then(
-    () => {
-      res.status(200).json({
-        message: 'Chocen sauce deleted!'
-      });
+  Sauce.findOne({_id: req.params.id}).then(
+    (sauce) => {
+      if(!sauce){
+        return res.status(404).json({
+          error: new Error('No such thing!')
+        });
+      }
+      if (sauce.userId !== req.auth.userId){
+        return res.status(400).json({
+          error: new Error('Unauthorized request!')
+        })
+      }
+      Sauce.deleteOne({ _id: req.parms.id }).then(
+        () => {
+          res.status(200).json({
+            message: 'Chocen sauce deleted!'
+          });
+        }
+      ).catch((error) => {
+        res.status(400).json({
+          error: error
+        });
+      }
+      );
     }
-  ).catch((error) => {
-    res.status(400).json({
-      error: error
-    });
-  }
-  );
+  )
 };
 
 exports.getAllSauces = (req, res, next) => {
